stripped out XMTP Context
