+++ components/Conversation/MessagesList.tsx
@@ -24,14 +24,14 @@ const formatDate = (d?: Date) =>
 const MessageTile = ({ message }: MessageTileProps): JSX.Element => (
   <div className="flex items-start mx-auto mb-4">
     <Avatar peerAddress={message.senderAddress as string} />
-    <div className="ml-2">
+    <div className="ml-2 max-w-[95%]">
       <div>
         <AddressPill address={message.senderAddress as string} />
         <span className="text-sm font-normal place-self-end text-n-300 text-md uppercase">
           {formatTime(message.sent)}
         </span>
       </div>
-      <span className="block text-md px-2 mt-2 text-black font-normal">
+      <span className="block text-md px-2 mt-2 text-black font-normal break-words">
         {message.error ? (
           `Error: ${message.error?.message}`
         ) : (
