+++ components/Layout.tsx
@@ -12,6 +12,7 @@ import { useCallback, useContext } from 'react'
 import { WalletContext } from '../contexts/wallet'
 import { useAppStore } from '../store/app'
 import useInitXmtpClient from '../hooks/useInitXmtpClient'
+import useListConversations from '../hooks/useListConversations'
 
 const NavigationColumnLayout: React.FC = ({ children }) => (
   <aside className="flex w-full md:w-84 flex-col flex-grow fixed inset-y-0">
@@ -67,6 +68,7 @@ const ConversationLayout: React.FC = ({ children }) => {
 const Layout: React.FC = ({ children }) => {
   const client = useAppStore((state) => state.client)
   const { initClient } = useInitXmtpClient()
+  useListConversations()
   const walletAddress = useAppStore((state) => state.address)
   const signer = useAppStore((state) => state.signer)
