--- contexts/xmtp.ts
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-export const XmtpContext = createContext({})
-
-export default XmtpContext
