+++ hooks/useInitXmtpClient.ts
@@ -1,12 +1,21 @@
 import { Client } from '@xmtp/xmtp-js'
 import { Signer } from 'ethers'
-import { useCallback } from 'react'
+import { useCallback, useEffect } from 'react'
 import { getEnv } from '../helpers'
 import { useAppStore } from '../store/app'
 
 const useInitXmtpClient = () => {
+  const signer = useAppStore((state) => state.signer)
   const client = useAppStore((state) => state.client)
   const setClient = useAppStore((state) => state.setClient)
+  const setConvoMessages = useAppStore((state) => state.setConvoMessages)
+  const setConversations = useAppStore((state) => state.setConversations)
+
+  const disconnect = () => {
+    setClient(undefined)
+    setConversations(new Map())
+    setConvoMessages(new Map())
+  }
 
   const initClient = useCallback(
     async (wallet: Signer) => {
@@ -21,6 +30,10 @@ const useInitXmtpClient = () => {
     },
     [client]
   )
+
+  useEffect(() => {
+    signer ? initClient(signer) : disconnect()
+  }, [signer])
 
   return {
     initClient,
