+++ hooks/useListConversations.tsx
@@ -1,13 +1,9 @@
 import { useEffect } from 'react'
-import { XmtpContext } from '../contexts/xmtp'
 import { useAppStore } from '../store/app'
-import useInitXmtpClient from '../hooks/useInitXmtpClient'
 
-export const XmtpProvider: React.FC = ({ children }) => {
+export const useListConversations = () => {
   const walletAddress = useAppStore((state) => state.address)
-  const signer = useAppStore((state) => state.signer)
   const client = useAppStore((state) => state.client)
-  const setClient = useAppStore((state) => state.setClient)
   const convoMessages = useAppStore((state) => state.convoMessages)
   const setConvoMessages = useAppStore((state) => state.setConvoMessages)
   const conversations = useAppStore((state) => state.conversations)
@@ -15,17 +11,6 @@ export const XmtpProvider: React.FC = ({ children }) => {
   const setLoadingConversations = useAppStore(
     (state) => state.setLoadingConversations
   )
-  const { initClient } = useInitXmtpClient()
-
-  const disconnect = () => {
-    setClient(undefined)
-    setConversations(new Map())
-    setConvoMessages(new Map())
-  }
-
-  useEffect(() => {
-    signer ? initClient(signer) : disconnect()
-  }, [signer])
 
   useEffect(() => {
     if (!client) return
@@ -65,9 +50,7 @@ export const XmtpProvider: React.FC = ({ children }) => {
     }
     listConversations()
     streamConversations()
-  }, [client])
-
-  return <XmtpContext.Provider value={{}}>{children}</XmtpContext.Provider>
+  }, [client, walletAddress])
 }
 
-export default XmtpProvider
+export default useListConversations
